requirements.txt
arxiv==2.1.3
beautifulsoup4==4.12.3
bs4==0.0.2
chromadb==0.5.5
faiss-cpu==1.8.0.post1
fastapi==0.103.0
groq==0.11.0
huggingface-hub==0.25.1
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langchain-ollama==0.2.0
langchain-openai==0.1.22
langchain-text-splitters==0.3.0
langchainhub==0.1.21
langserve==0.2.3
langsmith==0.1.132
matplotlib-inline==0.1.7
numpy==1.26.4
ollama==0.3.3
onnxruntime==1.19.2
openai==0.28.0
pandas==2.2.3
pillow==10.4.0
pydantic==2.7.4
pypdf==5.0.1
python-dotenv==1.0.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
SQLAlchemy==2.0.35
sse-starlette==2.1.3
starlette==0.27.0
streamlit==1.39.0
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.1
urllib3==2.2.3
uvicorn==0.22.0
wikipedia==1.4.0